backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
attrs==25.3.0
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
cryptography==44.0.2
cssselect2==0.8.0
deprecation==2.1.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
fonttools==4.57.0
frozenlist==1.5.0
gotrue==2.12.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.8
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.4.3
packaging==24.2
pillow==11.2.1
pluggy==1.5.0
postgrest==1.0.1
propcache==0.3.1
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pydyf==0.11.0
Pygments==2.19.1
PyJWT==2.10.1
pyphen==0.17.2
pytest==8.3.5
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
realtime==2.4.2
rich==14.0.0
rich-toolkit==0.14.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.0
supafunc==0.9.4
tinycss2==1.4.0
tinyhtml5==2.0.0
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.5
weasyprint==52.5
webencodings==0.5.1
websockets==14.2
yarl==1.19.0
zopfli==0.2.3.post1
mangum==0.19.0
